feat: 장바구니 개별 및 전체 선택 기능 구현

+++ app/src/main/java/woowacourse/shopping/presentation/cart/model/CartUiState.kt
@@ -1,7 +1,7 @@
 package woowacourse.shopping.presentation.cart.model
 
 data class CartUiState(
-    val totalPrice: Int = 0,
+    val totalPrice: Long = 0,
     val totalQuantity: Int = 0,
     val isSelectAll: Boolean = true,
     val page: Int = 0,
